Get rid of code which breaks Multi-ISO boot [Testing: issue1273]
Removed lots of code which is not needed any longer because grml2usb
can do a lot of the work.
Did some redesign to support multiple ISOs, see also bug report #1273.
Because of this changes it's no longer possible to overwrite the used
kernel in /etc/grml/terminalserver/config. This should not be needed
nowadays, anyway.

Fixed syntax error and wrong variable assignement

In default_config the variable MULTIBOOT was assigned wrong