Removed many old code and also unused variables
Removed lots of code which is not needed anylonger because grml2usb
can do a lot of the work.
Did some redesign to support multiple ISOs, see also bug report #1273.
Because of this changes it's no longer possible to overwrite the used
kernel in /etc/grml/terminalserver/config. This should not be needed
nowadays, anyway.